backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pdfplumber==0.11.4
jsonschema==4.22.0
python-dotenv==1.0.1
google-generativeai==0.7.2
